Add missing device arguments and move `hasBox` predicate
