+++ packages/alfa-dom/src/node/element/predicate.ts
@@ -1,5 +1,4 @@
 export * from "./predicate/has-attribute.js";
-export * from "./predicate/has-box.js";
 export * from "./predicate/has-display-size.js";
 export * from "./predicate/has-id.js";
 export * from "./predicate/has-input-type.js";
@@ -7,13 +6,13 @@ export * from "./predicate/has-name.js";
 export * from "./predicate/has-namespace.js";
 export * from "./predicate/has-tab-index.js";
 export * from "./predicate/has-unique-id.js";
+export * from "./predicate/is-actually-disabled.js";
 export * from "./predicate/is-browsing-context-container.js";
 export * from "./predicate/is-content.js";
-export * from "./predicate/is-actually-disabled.js";
 export * from "./predicate/is-document-element.js";
 export * from "./predicate/is-draggable.js";
 export * from "./predicate/is-editing-host.js";
 export * from "./predicate/is-fallback.js";
+export * from "./predicate/is-replaced.js";
 export * from "./predicate/is-scoped-to.js";
 export * from "./predicate/is-suggested-focusable.js";
-export * from "./predicate/is-replaced.js";
