+++ packages/alfa-dom/src/node/predicate.ts
@@ -1,3 +1,4 @@
+export * from "./predicate/has-box.js";
 export * from "./predicate/has-child.js";
 export * from "./predicate/has-descendant.js";
 export * from "./predicate/has-inclusive-descendant.js";
