+++ packages/alfa-dom/src/node/predicate/has-box.ts
@@ -1,7 +1,8 @@
 import type { Device } from "@siteimprove/alfa-device";
 import type { Predicate } from "@siteimprove/alfa-predicate";
 import type { Rectangle } from "@siteimprove/alfa-rectangle";
-import type { Element } from "../../element.js";
+import type { Element } from "../element.js";
+import type { Text } from "../text.js";
 
 /**
  * @public
@@ -9,6 +10,6 @@ import type { Element } from "../../element.js";
 export function hasBox(
   predicate: Predicate<Rectangle> = () => true,
   device: Device,
-): Predicate<Element> {
-  return (element) => element.getBoundingBox(device).some(predicate);
+): Predicate<Element | Text> {
+  return (node) => node.getBoundingBox(device).some(predicate);
 }
