+++ packages/alfa-dom/src/tsconfig.json
@@ -21,7 +21,6 @@
     "./node/element/input-type.ts",
     "./node/element/predicate.ts",
     "./node/element/predicate/has-attribute.ts",
-    "./node/element/predicate/has-box.ts",
     "./node/element/predicate/has-display-size.ts",
     "./node/element/predicate/has-id.ts",
     "./node/element/predicate/has-input-type.ts",
@@ -41,6 +40,7 @@
     "./node/element/predicate/is-suggested-focusable.ts",
     "./node/fragment.ts",
     "./node/predicate.ts",
+    "./node/predicate/has-box.ts",
     "./node/predicate/has-child.ts",
     "./node/predicate/has-descendant.ts",
     "./node/predicate/has-inclusive-descendant.ts",
